OUpdate build script

+++ buildandflash.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
 
 # Check if the script is run with an argument
-if [ $# -eq 0 ]; then
+if [ $# -gt 1 ]; then
     echo "Usage: $0 <device>"
     exit 1
+elif [ $# -gt 1 ]; then
+    sudo -v
 fi
 
 # Store the input argument in a variable
 device=$1
-
-sudo -v
 
 echo "Update flake.lock file..."
 nix flake lock --update-input ghaf
@@ -17,7 +17,7 @@ nix flake lock --update-input ghaf
 echo "Building..."
 nix build .# --impure
 
-if [ $? -eq 0 ]; then
+if [ $? -eq 0 ] && [[ -n "$device" ]]; then
     echo "Flashing..."
     sync; umount /media/$USER/*; sudo dd if=result/disk1.raw of=$device bs=1M status=progress conv=fsync
 fi
